check for even sides in focal() matrix
The non-integer maxDimension object is used in the creation of the matrix passed to the `focal()` function when creating the `lm.pts` object. Creating the `w` matrix using the `matrix()` function on non-integer values for the `nrow=` and `ncol=` arguments can result in a matrix with even dimensions, which throws a "w must have uneven sides" error when passed to the `raster::focal()` function (see line 74: https://github.com/cran/raster/blob/3e33743a12d2c8f407a4afc45dabc9ae9b1ae5fe/R/focal.R).

This patch uses the `ceiling()` function to ensure that `maxDimensions` is an integer, then increments it by 1 if it is an even value (assessed using `maxDimensions %% 2`).